pyproject.toml
[project]
name = "rekollect"
version = "2.0.0"
description = "Memory infrastructure for AI agents. Store, search, and recall anything."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "asyncpg>=0.30",
    "openai>=1.0",
    "pydantic-settings>=2.0",
    "bcrypt>=4.0",
    "graphiti-core @ git+https://github.com/290design/graphiti.git",
    "fastmcp>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-asyncio",
    "pytest-order",
    "httpx>=0.27",
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
